# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "HubGit Provider API",
    "version": "0.1.0",
    "description": "Provider-neutral contract implemented by HubGit mock and future Git host adapters."
  },
  "servers": [{ "url": "/", "description": "Current HubGit instance" }],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/api/v1/dashboard": {
      "get": { "tags": ["identity"], "operationId": "getDashboard" , "summary": "get dashboard", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }], "responses": { "200": { "description": "Viewer dashboard", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Dashboard" } } } }, "401": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/api/v1/notifications": {
      "get": { "tags": ["notifications"], "operationId": "listNotifications" , "summary": "list notifications", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "name": "unread", "in": "query", "schema": { "type": "boolean" } }], "responses": { "200": { "description": "Notifications", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NotificationPage" } } } } } },
      "patch": { "tags": ["notifications"], "operationId": "markNotificationsRead" , "summary": "mark notifications read", "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "ids": { "type": "array", "items": { "type": "string" } }, "all": { "type": "boolean" } } } } } }, "responses": { "204": { "description": "Notifications updated" } } }
    },
    "/api/v1/search": {
      "get": { "tags": ["search"], "operationId": "search" , "summary": "search", "security": [], "parameters": [{ "$ref": "#/components/parameters/QueryRequired" }, { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["repositories", "code", "commits", "issues", "pulls", "discussions", "users", "organizations"] } }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Search results", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchResultPage" } } } }, "422": { "$ref": "#/components/responses/Problem" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hubgit_session" }
    },
    "headers": {
...
      "NoStore": { "description": "Sensitive response must not be stored by shared caches", "schema": { "type": "string", "const": "no-store" } },
...
    },
    "parameters": {
...
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 30 } },
...
      "QueryRequired": { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 1 } },
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } },
...
    },
...
    "responses": {
      "Problem": { "description": "Problem details", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
...
    },
    "schemas": {
      "ProblemDetails": {
        "type": "object", "required": ["type", "title", "status", "code"],
        "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer" }, "detail": { "type": "string" }, "instance": { "type": "string", "format": "uri-reference" }, "code": { "type": "string", "examples": ["repository.not_found"] }, "fieldErrors": { "type": "array", "items": { "$ref": "#/components/schemas/FieldError" } }, "traceId": { "type": "string" } }
      },
      "FieldError": { "type": "object", "required": ["field", "code", "message"], "properties": { "field": { "type": "string" }, "code": { "type": "string" }, "message": { "type": "string" } } },
      "PageInfo": { "type": "object", "required": ["hasNextPage", "hasPreviousPage"], "properties": { "startCursor": { "type": ["string", "null"] }, "endCursor": { "type": ["string", "null"] }, "hasNextPage": { "type": "boolean" }, "hasPreviousPage": { "type": "boolean" } } },
      "ResourcePermissions": { "type": "object", "required": ["read", "write", "admin"], "properties": { "read": { "type": "boolean" }, "triage": { "type": "boolean", "default": false }, "write": { "type": "boolean" }, "maintain": { "type": "boolean", "default": false }, "admin": { "type": "boolean" } }, "additionalProperties": false },
...
      "Freshness": { "type": "object", "description": "Canonical synchronization and authorization metadata for data that may be served from an adapter cache.", "required": ["state", "lastSyncedAt", "lastAuthorizedAt", "provider"], "properties": { "state": { "type": "string", "enum": ["live", "refreshing", "stale", "offline"] }, "lastSyncedAt": { "type": ["string", "null"], "format": "date-time" }, "lastAuthorizedAt": { "type": ["string", "null"], "format": "date-time" }, "provider": { "type": "string" } }, "additionalProperties": false },
...
      "User": { "type": "object", "required": ["id", "kind", "username", "displayName", "avatarUrl", "profileUrl"], "properties": { "id": { "type": "string" }, "kind": { "const": "user" }, "username": { "type": "string" }, "displayName": { "type": "string" }, "avatarUrl": { "type": "string", "format": "uri-reference" }, "profileUrl": { "type": "string", "format": "uri-reference" }, "bio": { "type": ["string", "null"] }, "location": { "type": ["string", "null"] }, "followerCount": { "type": "integer", "minimum": 0 }, "followingCount": { "type": "integer", "minimum": 0 }, "createdAt": { "type": "string", "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } } },
...
      "Repository": {
        "type": "object", "required": ["id", "kind", "owner", "name", "fullName", "visibility", "defaultBranch", "empty", "archived", "permissions", "counts", "cloneUrls", "createdAt", "updatedAt"],
        "properties": { "id": { "type": "string" }, "kind": { "const": "repository" }, "owner": { "$ref": "#/components/schemas/RepositoryOwner" }, "name": { "type": "string" }, "fullName": { "type": "string" }, "description": { "type": ["string", "null"] }, "visibility": { "type": "string", "enum": ["public", "internal", "private"] }, "defaultBranch": { "type": "string" }, "empty": { "type": "boolean" }, "archived": { "type": "boolean" }, "fork": { "type": "boolean", "default": false }, "forkedFrom": { "type": ["string", "null"] }, "language": { "type": ["string", "null"] }, "license": { "type": ["string", "null"] }, "topics": { "type": "array", "items": { "type": "string" } }, "permissions": { "$ref": "#/components/schemas/ResourcePermissions" }, "counts": { "$ref": "#/components/schemas/RepositoryCounts" }, "cloneUrls": { "$ref": "#/components/schemas/CloneUrls" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }, "pushedAt": { "type": ["string", "null"], "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } }
      },
      "RepositoryOwner": { "type": "object", "required": ["id", "kind", "login", "avatarUrl"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "enum": ["user", "organization"] }, "login": { "type": "string" }, "avatarUrl": { "type": "string", "format": "uri-reference" } } },
      "RepositoryCounts": { "type": "object", "required": ["stars", "forks", "watchers", "issues", "pullRequests"], "properties": { "stars": { "type": "integer" }, "forks": { "type": "integer" }, "watchers": { "type": "integer" }, "issues": { "type": "integer" }, "pullRequests": { "type": "integer" } } },
      "CloneUrls": { "type": "object", "required": ["http", "ssh"], "properties": { "http": { "type": "string", "format": "uri-reference" }, "ssh": { "type": "string" } } },
...
      "RepositoryPage": { "type": "object", "required": ["items", "pageInfo"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Repository" } }, "pageInfo": { "$ref": "#/components/schemas/PageInfo" }, "totalCount": { "type": "integer" } } },
      "Dashboard": { "type": "object", "required": ["repositories", "activity"], "properties": { "repositories": { "$ref": "#/components/schemas/RepositoryPage" }, "activity": { "type": "array", "items": { "$ref": "#/components/schemas/TimelineEvent" } } } },
...
      "Label": { "type": "object", "required": ["id", "name", "color", "description"], "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "color": { "type": "string", "pattern": "^[0-9a-fA-F]{6}$" }, "description": { "type": "string" } } },
...
      "Milestone": { "type": "object", "required": ["id", "number", "title", "state", "openIssueCount", "closedIssueCount"], "properties": { "id": { "type": "string" }, "number": { "type": "integer" }, "title": { "type": "string" }, "description": { "type": ["string", "null"] }, "state": { "type": "string", "enum": ["open", "closed"] }, "dueAt": { "type": ["string", "null"], "format": "date-time" }, "openIssueCount": { "type": "integer" }, "closedIssueCount": { "type": "integer" } } },
...
      "Issue": { "type": "object", "required": ["id", "kind", "number", "title", "body", "state", "author", "labels", "assignees", "commentCount", "permissions", "createdAt", "updatedAt"], "properties": { "id": { "type": "string" }, "kind": { "const": "issue" }, "number": { "type": "integer" }, "title": { "type": "string" }, "body": { "type": "string" }, "state": { "type": "string", "enum": ["open", "closed"] }, "stateReason": { "type": ["string", "null"], "enum": ["completed", "not_planned", "reopened", null] }, "author": { "$ref": "#/components/schemas/User" }, "labels": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } }, "assignees": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }, "milestone": { "oneOf": [{ "$ref": "#/components/schemas/Milestone" }, { "type": "null" }] }, "commentCount": { "type": "integer" }, "permissions": { "$ref": "#/components/schemas/ResourcePermissions" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }, "closedAt": { "type": ["string", "null"], "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } } },
...
      "PullRequest": { "type": "object", "required": ["id", "kind", "number", "title", "body", "state", "draft", "author", "base", "head", "mergeability", "reviewDecision", "checks", "permissions", "createdAt", "updatedAt"], "properties": { "id": { "type": "string" }, "kind": { "const": "pullRequest" }, "number": { "type": "integer" }, "title": { "type": "string" }, "body": { "type": "string" }, "state": { "type": "string", "enum": ["open", "closed", "merged"] }, "draft": { "type": "boolean" }, "author": { "$ref": "#/components/schemas/User" }, "base": { "$ref": "#/components/schemas/PullRequestRef" }, "head": { "$ref": "#/components/schemas/PullRequestRef" }, "mergeability": { "type": "string", "enum": ["mergeable", "conflicting", "checking", "blocked", "unknown"] }, "reviewDecision": { "type": "string", "enum": ["approved", "changes_requested", "review_required", "none"] }, "checks": { "$ref": "#/components/schemas/CheckSummary" }, "labels": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } }, "permissions": { "$ref": "#/components/schemas/ResourcePermissions" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }, "mergedAt": { "type": ["string", "null"], "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } } },
      "PullRequestRef": { "type": "object", "required": ["repository", "ref", "sha"], "properties": { "repository": { "type": "string" }, "ref": { "type": "string" }, "sha": { "type": "string" } } },
...
      "CheckSummary": { "type": "object", "required": ["total", "queued", "inProgress", "successful", "failed"], "properties": { "total": { "type": "integer" }, "queued": { "type": "integer" }, "inProgress": { "type": "integer" }, "successful": { "type": "integer" }, "failed": { "type": "integer" } } },
...
      "Notification": { "type": "object", "required": ["id", "reason", "unread", "subject", "repository", "updatedAt"], "properties": { "id": { "type": "string" }, "reason": { "type": "string", "enum": ["assign", "author", "comment", "mention", "review_requested", "security_alert", "subscribed", "team_mention"] }, "unread": { "type": "boolean" }, "subject": { "$ref": "#/components/schemas/SubjectReference" }, "repository": { "$ref": "#/components/schemas/Repository" }, "updatedAt": { "type": "string", "format": "date-time" }, "lastReadAt": { "type": ["string", "null"], "format": "date-time" } } },
      "SubjectReference": { "type": "object", "required": ["kind", "id", "title", "url"], "properties": { "kind": { "type": "string", "enum": ["issue", "pullRequest", "discussion", "release", "workflowRun", "repository"] }, "id": { "type": "string" }, "title": { "type": "string" }, "url": { "type": "string", "format": "uri-reference" } } },
      "NotificationPage": { "type": "object", "required": ["items", "pageInfo"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } }, "pageInfo": { "$ref": "#/components/schemas/PageInfo" } } },
      "SearchResult": { "oneOf": [{ "$ref": "#/components/schemas/RepositorySearchResult" }, { "$ref": "#/components/schemas/UserSearchResult" }, { "$ref": "#/components/schemas/IssueSearchResult" }, { "$ref": "#/components/schemas/CodeSearchResult" }], "discriminator": { "propertyName": "kind" } },
      "RepositorySearchResult": { "type": "object", "required": ["kind", "score", "repository"], "properties": { "kind": { "const": "repository" }, "score": { "type": "number" }, "repository": { "$ref": "#/components/schemas/Repository" } } },
      "UserSearchResult": { "type": "object", "required": ["kind", "score", "user"], "properties": { "kind": { "const": "user" }, "score": { "type": "number" }, "user": { "$ref": "#/components/schemas/User" } } },
      "IssueSearchResult": { "type": "object", "required": ["kind", "score", "repository", "issue"], "properties": { "kind": { "type": "string", "enum": ["issue", "pullRequest"] }, "score": { "type": "number" }, "repository": { "type": "string" }, "issue": { "oneOf": [{ "$ref": "#/components/schemas/Issue" }, { "$ref": "#/components/schemas/PullRequest" }] } } },
      "CodeSearchResult": { "type": "object", "required": ["kind", "score", "repository", "path", "matches"], "properties": { "kind": { "const": "code" }, "score": { "type": "number" }, "repository": { "type": "string" }, "path": { "type": "string" }, "matches": { "type": "array", "items": { "type": "object", "required": ["line", "fragment"], "properties": { "line": { "type": "integer" }, "fragment": { "type": "string" } } } } } },
      "SearchResultPage": { "type": "object", "required": ["items", "pageInfo"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/SearchResult" } }, "pageInfo": { "$ref": "#/components/schemas/PageInfo" }, "totalCount": { "type": "integer" } } },
      "TimelineEvent": { "type": "object", "required": ["id", "kind", "actor", "createdAt"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "enum": ["commit", "create", "comment", "stateChange", "assignment", "label", "review", "merge", "release"] }, "actor": { "$ref": "#/components/schemas/User" }, "subject": { "$ref": "#/components/schemas/SubjectReference" }, "createdAt": { "type": "string", "format": "date-time" }, "metadata": { "type": "object", "additionalProperties": true } } }
    }
  }
}